+++ R/utils.R
@@ -1,11 +1,11 @@
 #' From a set of intervals, find which interval values belong to
 #'
-#' This utility function is a replacement for \code{findInterval} that works even when the set of
-#' intervals is discontinuous. It wraps \code{\link[intervals]{which_nearest}} from the
-#' "intervals" package.
-#' @param x A set of numeric values
+#' Returns which interval (if any) each number in a vector belongs to, given a set of
+#' user-defined intervals. Intervals can be specified using either two-column matrices or
+#' \code{\link[intervals:Intervals-class]{Intervals}} objects from the \code{intervals} package.
+#' @param x A vector of numeric values
 #' @param Intv A two-column matrix or an object of class Intervals
-#' @return For each value in x: if x[i] in in the set of intervals, the index of the corresponding
+#' @return For each value in x: if x[i] is in the set of intervals, the index of the corresponding
 #'   interval(s), NA if no interval contains x[i]
 #' @seealso \%In\%
 #' @examples
@@ -33,10 +33,17 @@ whichInterval <- function(x, Intv) {
     wn$which_nearest
 }
 
+#' @rdname whichInterval
+#' @export
+which_interval <- whichInterval  # Alias for keeping with tidyverse-style naming conventions
+
 #' Find if value belongs to a set of intervals
 #'
-#' Wrapper around \code{\link[intervals]{distance_to_nearest}} from the "intervals" package.
-#' @param x A set of numeric values
+#' Returns whether numeric values on the left-hand side of the operator fall within any of the
+#' specified intervals on the right-hand side. Intervals can be specified using either two-column
+#' matrices or \code{\link[intervals:Intervals-class]{Intervals}} objects from the
+#' \code{intervals} package.
+#' @param x A vector of numeric values
 #' @param Intv A set of intervals, defined by a two-column matrix of endpoints or an Intervals
 #'   object
 #' @return A vector of logicals, which are true if x[i] belongs to any of the intervals in the set.
@@ -54,3 +61,7 @@ whichInterval <- function(x, Intv) {
     }
     distance_to_nearest(x, Intv) == 0
 }
+
+#' @rdname grapes-In-grapes
+#' @export
+`%within%` <- `%In%`  # Alias for keeping with tidyverse-style naming conventions